Split draw.rs into modular files. Some refactorings.
